+++ src/main/java/lotto/model/wrapper/LottoNumber.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.wrapper;
 
 import java.util.Objects;
