+++ src/test/java/LottoTests/LottoStoreTests.java
@@ -2,7 +2,7 @@ package LottoTests;
 
 import lotto.model.LottoPurchaseTickets;
 import lotto.model.LottoStore;
-import lotto.model.Payment;
+import lotto.model.wrapper.Payment;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
